Moved a config struct to the right package.
AuthConfig reads a file, so it should go into config/
