+++ model/auth.go
@@ -10,18 +10,6 @@ type AuthError struct {
 	} `json:"errors"`
 }
 
-// Located at /etc/session-counter/auth.yaml
-type AuthConfig struct {
-	Directus struct {
-		Token string `yaml:"token"`
-		User  string `yaml:"username"`
-	} `yaml:"directus"`
-	Reval struct {
-		Token string `yaml:"token"`
-		User  string `yaml:"username"`
-	} `yaml:"reval"`
-}
-
 type Auth struct {
 	Token string `yaml:"token"`
 	User  string `yaml:"username"`
